rename wsnet2-bot
