+++ server/cmd/wsnet2-bot/cmd/root.go
