+++ server/cmd/wsnet2-bot/cmd/scenario.go
