+++ server/cmd/wsnet2-bot/main.go
@@ -1,6 +1,6 @@
 package main
 
-import "wsnet2/cmd/wsnet2-bot2/cmd"
+import "wsnet2/cmd/wsnet2-bot/cmd"
 
 func main() {
 	cmd.Execute()
